Ensure only modified files be processed the second time

+++ pydgutils/__init__.py
@@ -38,6 +38,25 @@ def __copy_tree(src_dir, dest_dir):
         else:
             shutil.copy2(from_path, to_path)
 
+def remove_temporary_directories(base_dir):
+    # The 'build' and 'dist' folder sometimes will not update! So we need to
+    # remove them all !
+    leave_dir = os.path.realpath(os.path.normpath(os.path.join(base_dir, _PROCESSED_DIR)))
+    root_dir = os.path.join(base_dir, "build")
+    if not os.path.exists(root_dir):
+        return
+
+    files = os.listdir(root_dir)
+    for afile in files:
+        afile_path = os.path.realpath(os.path.normpath(os.path.join(root_dir, afile)))
+        if os.path.isdir(afile_path):
+            if not os.path.samefile(afile_path, leave_dir):
+                shutil.rmtree(afile_path, ignore_errors=True)
+        else:
+            os.remove(afile_path)
+
+    shutil.rmtree(os.path.join(base_dir, 'dist'), ignore_errors=True)
+
 def downgrade_files(files):
     # Check and prepare 3to2 module.
     try:
@@ -65,6 +84,12 @@ def downgrade_files(files):
         options += files
 
         lib3to2_main("lib3to2.fixes", options)
+
+def is_path_in_filters(apath, filters):
+    for afilter in filters:
+        if fnmatch.fnmatch(apath, afilter):
+            return True
+    return False
 
 def process(base_dir=os.curdir):
     """
@@ -97,10 +122,7 @@ def process(base_dir=os.curdir):
     if not os.path.exists(source_path):
         raise Dir
 
-    # The 'build' and 'dist' folder sometimes will not update! So we need to
-    # remove them all !
-    shutil.rmtree(os.path.join(base_dir, 'build'), ignore_errors=True)
-    shutil.rmtree(os.path.join(base_dir, 'dist'), ignore_errors=True)
+    remove_temporary_directories(base_dir)
 
     # Remove all unused directories
     directories = []
@@ -135,6 +157,8 @@ def process(base_dir=os.curdir):
         cmp_result = filecmp.dircmp(source_path, destination_path)
         dirs.append(cmp_result)
 
+        dir_filters = ["__pycache__"]
+
         while len(dirs) > 0:
 
             # Get the last one compare result
@@ -166,19 +190,27 @@ def process(base_dir=os.curdir):
 
                 if os.path.isdir(left_file_path):
                     __copy_tree(left_file_path, right_file_path)
-                    files.append(right_file_path)
+                    if not is_path_in_filters(os.path.basename(right_file_path), dir_filters):
+                        files.append(right_file_path)
                     continue
 
                 if not fnmatch.fnmatch(file_name, "*.py"):
                     continue
 
-                try:
-                    os.remove(right_file_path)
-                except:
-                    pass
+                left_file_mtime = os.stat(left_file_path).st_mtime
+                right_file_mtime = 0
 
-                shutil.copy2(left_file_path, right_file_path)
-                files.append(right_file_path)
+                if os.path.exists(right_file_path):
+                    right_file_mtime = os.stat(right_file_path).st_mtime
+
+                if left_file_mtime > right_file_mtime:
+                    try:
+                        os.remove(right_file_path)
+                    except:
+                        pass
+
+                    shutil.copy2(left_file_path, right_file_path)
+                    files.append(right_file_path)
 
         if len(files) > 0:
             downgrade_files(files)
